+++ levels/revert.rb
@@ -1,5 +1,12 @@
 difficulty 4
-description "You have committed several times but want to undo the middle commit. All commits have been pushed, so you can't change existing history."
+# description "You have committed several times but want to undo the middle commit. All commits have been pushed, so you can't change existing history."
+text = <<~TEXT
+  Sie haben mehrere Commits durchgeführt, möchten aber den mittleren
+  Commit rückgängig machen. Alle Commits wurden bereits auf das
+  Remote Repository übertragen (push), so dass Sie die bestehende
+  Historie nicht ändern können.
+TEXT
+description text
 
 setup do
   repo.init
@@ -19,12 +26,13 @@ end
 
 solution do
   valid = false
-  commit_messages = repo.commits.map(&:message)
-  valid = true if repo.commits.length > 3 &&
+  commit_messages = repo.commits(repo.head.name).map(&:message)
+  valid = true if repo.commits(repo.head.name).length > 3 &&
     commit_messages.any? { |e| e =~ /(Revert )?"Bad commit"/ }
   valid
 end
 
 hint do
-  puts "Try the revert command."
+  # puts "Try the revert command."
+  puts "Versuchen Sie den Befehl `git revert`."
 end