+++ CHANGELOG.md
@@ -58,9 +58,13 @@ Der \<Scope> kann alles sein, was den Ort der Veränderung angibt.
 - German translation of level 41-50  
   rebase_onto, repack, cherry-pick, grep, rename_commit,  
   squash, merge_squash, reorder, bisect, stage_lines
+- German translation of level 51-56  
+  find_old_branch, revert, restore, conflict, submodule,  
+  contribute
 
 ## 2022-08-04
 - feat: German translation of level 41-50
+- feat: German translation of level 51-56
 
 ## 2022-08-03
 - feat: German translation of level 31-40